miner-fs: Remove conditional monitor initialization on folder count
This used to be a measure for tracker_monitor_add() being constly
for inotify, at a time that inotify was not separated to a different
thread/context.

Nowadays, fanotify is our first choice, and even for inotify users
the slow bits are moved to a separate thread and don't affect everything
else. It seems we can drop this check and add monitors during
initialization right away.