tracker-miner-fs: Keep file path instead of GFile in fanotify monitor
Since we tend to potentially keep a high number of directory monitors
around, try to avoid dependency on GFile in order to spend too much
memory usage on needless payload, and stick to the minimum, file paths.

This avoids a small piece of string (file vs uri), and a GObject
wrapping around it, for every monitored directory. This is expected
to make idle tracker-miner-fs slimmer in memory, since most of memory
usage at that stage comes from TrackerMonitor.